feat(schema): money is bigint; public listings stop naming their owner
First cycle of porting Domain Schema v1 into the existing stack.

Every monetary column was int4 — a ceiling of 21 474 836,47 zł. Enough for a
month's rent, not for a sale price, and SALE is in scope. A 25 mln zł listing
would not have been rejected cleanly: the insert raised "integer out of range"
inside the transaction and the person listing it got a 500. Changed now because
the ALTER rewrites each table and the rewrite only gets more expensive. The
ledger's append-only triggers are unaffected: they fire on row UPDATE/DELETE,
and a type change is a table rewrite by the owner.

The guard reads the live catalogue rather than a list of columns: anything named
amount, fee, *_amount, *_fee or *_minor must be bigint, and no floating-point
column may exist anywhere. A new table with a price is covered the day its
migration runs. The naming convention is the contract, and its edges are tested
both ways so nobody loosens it until it matches nothing.

ClassifiedOut no longer carries owner_id. A stable account id on every public
listing let anyone join one person's whole portfolio together — addresses by
district included — without signing in.

Three mutants (two columns left int4, owner_id put back), all caught.
410 tests on Postgres.

+++ backend/app/modules/booking/models.py
@@ -1,7 +1,7 @@
 from datetime import date, datetime, timezone
 from uuid import uuid4
 
-from sqlalchemy import Date, DateTime, ForeignKey, Integer, String, UniqueConstraint
+from sqlalchemy import BigInteger, Date, DateTime, ForeignKey, Integer, String, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.core.config import settings
@@ -30,7 +30,7 @@ class Booking(Base):
     # pending -> confirmed -> completed | cancelled | expired (BK-01)
     #   expired = unpaid past its deadline; distinct from a user cancellation
     status: Mapped[str] = mapped_column(String(16), default="pending", index=True)
-    total_amount: Mapped[int] = mapped_column(Integer)  # minor units, ADR-0002
+    total_amount: Mapped[int] = mapped_column(BigInteger)  # minor units, ADR-0002
     currency: Mapped[str] = mapped_column(String(3))
     # The commission rate agreed AT BOOKING TIME, in basis points. Stamped here
     # rather than read from settings at payout: the rate is an admin-editable
